Convert objects, random, utils services to TS
